feature/logging user with google oauth to enable uploading content
This commit adds functionality to "Login" button in the UI.
After clicking on this button, Google OAuth window pops up and user may "login" for their account.
This means they allow the plugin to upload a YouTube video on their behalf.

Login function starts a listening server socket located in `OAuthReceiver.h` file.
Function creates a pointer to an instance of OAuthReceiver class and runs it on another thread.

It also assigns a callback that will be run when Google sends the authentication code.
This code is then catched and stored to `googleAuthCode` variable and will be used with upload function.
Lastly, it launches OAuth link in the browser.

Instance of OAuthReceiver class creates a server on `localhost:8080` and waits for Google OAuth to redirect user to this host.
When it happens, the instance tries to parse the authentication code provided by Google.
When it succeeds, callback is executed with authentication code as a parameter.

After any condition user is redirected to API's screen saying if logging operation was sucessful or not.

At the end, `googleAuthCode` contains authentication code that is necessary to obtain access and refresh token.