feature/refresh token based logging to not open the browser every time
Add `checkForRefreshToken` function that checks if file containing data necessary to login was created in APPDATA folder.

Update `login` function to handle it when user is clicking on Login button again after logging in. This stops them from crashing the plugin when opening another thread when one is running.
As of now, the user is only available to login once, if they want to chage selected account they must delete the plugin and reopen it

Add function that creates file with refresh token data, which will be used in upload function