feature/add `change account` button to allow for choosing different account
When user logged in on an account the refresh_token.txt file is created. Every time when user want to log in, data from this file is read and the same account as the previous time is choosen.

This button deletes this file and launches login function to allow user to choose different account.